add c6 and c1-demotion enable/disable to config command
